Refactor: 1:1문의 저장 및 목록 조회 로직 사용자/공급자 통합
