+++ src/main/resources/templates/contact/contact-list.html
@@ -17,6 +17,9 @@
 <div sec:authorize="hasAnyRole('ROLE_USER')">
     <div th:replace="~{fragments/header-user :: header}"></div>
 </div>
+<div sec:authorize="hasAnyRole('ROLE_OWNER')">
+    <div th:replace="~{fragments/header-owner :: header}"></div>
+</div>
 <body>
 <input type="hidden" th:name="${_csrf.parameterName}" th:value="${_csrf.token}"/>
 <main id="main">
@@ -55,8 +58,8 @@
                                                 <option value="3" th:selected="${request.searchName == '3'}">문의 취소</option>
                                             </select>
                                             <select id="answerYn" name="searchName" style="display: none;" aria-label="답변 여부">
-                                                <option value="Y" th:selected="${request.searchName == 'Y'}">문의 대기</option>
-                                                <option value="N" th:selected="${request.searchName == 'N'}">검토 중</option>
+                                                <option value="Y" th:selected="${request.searchName == 'Y'}">문의 완료</option>
+                                                <option value="N" th:selected="${request.searchName == 'N'}">문의 대기</option>
                                             </select>
                                             <button id="searchButton" type="button" class="btn btn-dark">검색</button>
                                             <button id="resetBtn" type="button" class="btn resetBtn-dark">초기화</button>
@@ -108,21 +111,21 @@
                     <th:block th:if="${request.totalCount > 0}">
                         <ul>
                             <li th:if="${request.startPageNo > 1}">
-                                <a th:href="@{/contact/user/list(pageNo=${request.firstPageNo})}">&laquo;</a>
+                                <a th:href="@{/contact/list(pageNo=${request.firstPageNo})}">&laquo;</a>
                             </li>
                             <li th:if="${request.pageNo > request.startPageNo}">
-                                <a th:href="@{/contact/user/list(pageNo=${request.prevPageNo})}">&lt;</a>
+                                <a th:href="@{/contact/list(pageNo=${request.prevPageNo})}">&lt;</a>
                             </li>
                             <th:block th:each="i : ${#numbers.sequence(request.startPageNo, request.endPageNo)}">
                                 <li th:classappend="${i == request.pageNo} ? 'active'">
-                                    <a th:href="@{/contact/user/list(pageNo=${i})}" th:text="${i}"></a>
+                                    <a th:href="@{/contact/list(pageNo=${i})}" th:text="${i}"></a>
                                 </li>
                             </th:block>
                             <li th:if="${request.pageNo < request.endPageNo}">
-                                <a th:href="@{/contact/user/list(pageNo=${request.nextPageNo})}">&gt;</a>
+                                <a th:href="@{/contact/list(pageNo=${request.nextPageNo})}">&gt;</a>
                             </li>
                             <li th:if="${request.endPageNo < request.finalPageNo}">
-                                <a th:href="@{/contact/user/list(pageNo=${request.finalPageNo})}">&raquo;</a>
+                                <a th:href="@{/contact/list(pageNo=${request.finalPageNo})}">&raquo;</a>
                             </li>
                         </ul>
                     </th:block>
@@ -185,15 +188,15 @@
                 searchValue = searchName.value.trim();
             }
 
-            window.location.href = `/contact/user/list?pageNo=${pageNo}&searchType=${searchType.value}&searchName=${searchValue}`;
+            window.location.href = `/contact/list?pageNo=${pageNo}&searchType=${searchType.value}&searchName=${searchValue}`;
         });
 
         contactWriteBtn.addEventListener("click", function () {
-           window.location.href = "/contact/user/write";
+           window.location.href = "/contact/write";
         });
 
         resetButton.addEventListener("click", function () {
-            window.location.href = "/contact/user/list";
+            window.location.href = "/contact/list";
         });
     });
 </script>
