+++ src/main/resources/templates/contact/contact-write.html
@@ -17,6 +17,9 @@
 <div sec:authorize="hasAnyRole('ROLE_USER')">
     <div th:replace="~{fragments/header-user :: header}"></div>
 </div>
+<div sec:authorize="hasAnyRole('ROLE_OWNER')">
+    <div th:replace="~{fragments/header-owner :: header}"></div>
+</div>
 <main id="main">
     <div class="breadcrumbs d-flex align-items-center" style="background-image: url('/assets/img/breadcrumbs-bg.jpg');">
     </div>
@@ -57,7 +60,7 @@
         const csrfToken = csrfTokenElement ? csrfTokenElement.value : null;
 
         backListButton.addEventListener("click", function () {
-            window.location.href = "/contact/user/list";
+            window.location.href = "/contact/list";
         });
 
         document.getElementById("writeForm").addEventListener("submit", function (event){
@@ -85,7 +88,7 @@
                 formData.append("contactFiles", fileList[i].file);
             }
 
-            fetch("/api/user/contact/save", {
+            fetch("/api/contact/save", {
                 method: "POST",
                 headers: {
                     "X-CSRF-TOKEN": csrfToken,
@@ -98,7 +101,7 @@
 
                     if (response.status === 200) {
                         alert(MSG.CONTACT_SAVE_SUCCESS);
-                        window.location.href = "/contact/user/list";
+                        window.location.href = "/contact/list";
                     } else {
                         alert(responseData.errorMsg);
                     }
