+++ src/main/resources/templates/fragments/header-user.html
@@ -28,7 +28,7 @@
                         <li class="dropdown"><a href="/board/list?page=1"><span>고객 문의</span> <i
                                 class="bi bi-chevron-down dropdown-indicator"></i></a>
                             <ul>
-                                <li><a href="/contact/user/list">1:1문의</a></li>
+                                <li><a href="/contact/list">1:1문의</a></li>
                                 <li><a href="/notice/list">공지사항</a></li>
                                 <li><a href="">FAQ</a></li>
                                 <li><a href="/board/list">자유게시판</a></li>
